==> Rmd/01_comparison_MAR.Rmd <==
---
title: "01_comparison_MAR"
output: html_document
editor_options: 
  markdown: 
    wrap: 72
---

```{r, setup}
# on fait en sorte que le wd soit bon aussi quand on exécute chunk by chunk
if (interactive()) {
  setwd(rprojroot::find_rstudio_root_file())
}
getwd()
```

# Chargement des libraries

```{r, results='hide', warnings = FALSE, message= FALSE}
library(ggplot2)
library(dplyr)
library(tidyr)
library(lubridate)
library(openair)
library(foehnix)
library(metR)
library(ncdf4)
library(circular)
library(patchwork)
# paramètres pour l'exécution du code
savefig = TRUE
getwd()
```

# Chargement des données

```{r}
# chargement des données
df_mf <- readRDS("./data/processed/MAR.rds")
df_ERA5 <- readRDS("./data/processed/ERA5_dataset_MAR.rds")
df_ERA5land <- readRDS("./data/processed/ERA5land_df_435_52.rds")
```

```{r}
figpath = ("./outputs/figures/comparison_MAR/")
```

```{r}
# on prend que les données de 2022 pour les df, jusqu'au 31 décembre 2022 inclus
df_ERA5 <- df_ERA5 %>%
  select(-c(datejulian, flag_dn, hour, month, year, saison, time))
df_ERA5 <- df_ERA5 %>%
  filter(date >= as.POSIXct("2022-01-01", tz = "UTC") & date < as.POSIXct("2023-01-01", tz = "UTC")) %>%
  rename_with(~ paste0(.x, "_era5"), .cols = -date) 

df_ERA5land <- df_ERA5land %>%
  select(-c(datejulian, flag_dn, time, saison))
df_ERA5land <- df_ERA5land %>%
  filter(date >= as.POSIXct("2022-01-01", tz = "UTC") & date < as.POSIXct("2023-01-01", tz = "UTC")) %>%
  
  rename_with(~ paste0(.x, "_era5land"), .cols = -date)

df_mf <- df_mf %>%
  select(-c(NUM_POSTE, NOM_USUEL, flag_dn, saison))
df_mf <- df_mf %>%
  filter(date >= as.POSIXct("2022-01-01", tz = "UTC") & date < as.POSIXct("2023-01-01", tz = "UTC")) %>%
  rename_with(~ paste0(.x, "_mf"), .cols = -date)

# on joint les toirs
df <- df_mf %>%
  left_join(df_ERA5, by = "date") %>%
  left_join(df_ERA5land, by = "date")

# on passe GLO_mf de J/cm² à J/m² puis à des W/m2
df$GLO_mf <- df$GLO_mf * 10000 /3600
# on renomme
df <- df %>%
  rename(
    ssrd_mf = GLO_mf,
    ws_mf = FF_mf,
    wd_mf = DD_mf,
    t2m_mf = T_mf,
    rh_mf = U_mf,
    totalprecipitation_mf = RR1_mf,
    sp_mf = PSTAT_mf
  )
```
# Stats
```{r}
# on calcule les rmse et bias par rapport à météofrance comme ref à chaque fois, pour era5 et era5 land
rmse_ws_era5 <- sqrt(mean((df$ws_mf - df$ws_era5)^2, na.rm = TRUE))
print(paste("RMSE WS ERA5:", round(rmse_ws_era5, 2)))
bias_ws_era5 <- mean(df$ws_mf - df$ws_era5, na.rm = TRUE)
print(paste("Bias WS ERA5:", round(bias_ws_era5, 2)))
rmse_ws_era5land <- sqrt(mean((df$ws_mf - df$ws_era5land)^2, na.rm = TRUE))
print(paste("RMSE WS ERA5-Land:", round(rmse_ws_era5land, 2)))
bias_ws_era5land <- mean(df$ws_mf - df$ws_era5land, na.rm = TRUE)
print(paste("Bias WS ERA5-Land:", round(bias_ws_era5land, 2)))
rmse_wd_era5 <- sqrt(mean((df$wd_mf - df$wd_era5)^2, na.rm = TRUE))
print(paste("RMSE WD ERA5:", round(rmse_wd_era5, 2)))
bias_wd_era5 <- mean(df$wd_mf - df$wd_era5, na.rm = TRUE)
print(paste("Bias WD ERA5:", round(bias_wd_era5, 2)))
rmse_t_era5 <- sqrt(mean((df$t2m_mf - df$t2m_era5)^2, na.rm = TRUE))
print(paste("RMSE T ERA5:", round(rmse_t_era5, 2)))
bias_t_era5 <- mean(df$t2m_mf - df$t2m_era5, na.rm = TRUE)
print(paste("Bias T ERA5:", round(bias_t_era5, 2)))
rmse_t_era5land <- sqrt(mean((df$t2m_mf - df$t2m_era5land)^2, na.rm = TRUE))
print(paste("RMSE T ERA5-Land:", round(rmse_t_era5land, 2)))
bias_t_era5land <- mean(df$t2m_mf - df$t2m_era5land, na.rm = TRUE)
print(paste("Bias T ERA5-Land:", round(bias_t_era5land, 2)))
rmse_ssrd_era5land <- sqrt(mean((df$ssrd_mf - df$ssrd_era5land)^2, na.rm = TRUE))
print(paste("RMSE SSRD ERA5-Land:", round(rmse_ssrd_era5land, 2)))



summary_df <- df %>%
  select(-date) %>%
  sapply(function(x) c(mean = mean(x, na.rm = TRUE),
                       sd = sd(x, na.rm = TRUE),
                       min = min(x, na.rm = TRUE),
                       max = max(x, na.rm = TRUE))) %>%
  t() %>%
  as.data.frame() %>%
  tibble::rownames_to_column("variable")
```
# Comparaison des données pour le vent

```{r}
# on prend 3 jours pour faire des graphiques
df_3j <- df %>%
  filter(date >= as.POSIXct("2022-01-10", tz = "UTC") & date < as.POSIXct("2022-01-14", tz = "UTC"))
# on crée une colonne avec les heures dsepuis le début du df
df_3j$heures <- as.numeric(difftime(df_3j$date, min(df_3j$date), units = "hours"))
```

```{r}
# plot as Menut
df_3j_long <- data.frame(
  heures = rep(df_3j$heures, 3),
  u = c(df_3j$windu_mf, df_3j$windu_era5, df_3j$windu_era5land),
  v = c(df_3j$windv_mf, df_3j$windv_era5, df_3j$windv_era5land),
  source = rep(c("Météo France", "ERA5", "ERA5-Land"), each = nrow(df_3j))
)

xlim_common <- range(df_3j$heures, na.rm = TRUE)
breaks_major <- seq(0, max(df_3j$heures), by = 24)
breaks_minor <- seq(0, max(df_3j$heures), by = 6)

# Flèche d'échelle uniquement pour Météo France
ref_arrow <- data.frame(
  heures = xlim_common[1] + 3,
  v = -9,
  u = 5,
  vv = 0,
  source = "Météo France",
  label = "5 m/s",
  stringsAsFactors = FALSE
)

ggplot(df_3j_long, aes(x = heures, y = 0)) +
  geom_segment(aes(xend = heures + u, yend = v), arrow = arrow(length = unit(0.15, "cm"))) +
  facet_wrap(~source, ncol = 1, strip.position = "left") +
  scale_x_continuous(
    name = NULL,
    limits = xlim_common,
    breaks = breaks_major,
    minor_breaks = breaks_minor,
    labels = function(x) format(min(df_3j$date) + x * 3600,  "%Y-%m-%d %H:%M")
  ) +
  coord_fixed() +
  theme_minimal() +
  theme(
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank()
  ) +
  geom_segment(
    data = ref_arrow,
    aes(x = heures, y = v, xend = heures + u, yend = v + vv),
    inherit.aes = FALSE,
    arrow = arrow(length = unit(0.15, "cm")),
    color = "red"
  ) +
  geom_text(
    data = ref_arrow,
    aes(x = heures + u + 1, y = v + vv, label = label),
    inherit.aes = FALSE,
    hjust = 0,
    vjust = 0.5,
    size = 3
  )

```

```{r, eval = savefig}
ggsave(
  filename = paste0(figpath, "comparison_MAR_wind_3j_big.png"),
  width = 18, height = 10, units = "cm", dpi = 300, bg = "white"
)
```

```{r}
ggplot(df_3j_long, aes(x = heures, y = 0, color = source)) +
  geom_segment(aes(xend = heures + u, yend = v), arrow = arrow(length = unit(0.15, "cm"))) +
  scale_x_continuous(
    name = NULL,
    limits = xlim_common,
    breaks = breaks_major,
    minor_breaks = breaks_minor,
    labels = function(x) format(min(df_3j$date) + x * 3600,  "%Y-%m-%d %H:%M")
  ) +
  coord_fixed() +
  theme_minimal() +
  theme(
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank()
  ) +
  geom_segment(
    data = ref_arrow,
    aes(x = heures, y = v, xend = heures + u, yend = v + vv),
    inherit.aes = FALSE,
    arrow = arrow(length = unit(0.15, "cm")),
    color = "black"
  ) +
  geom_text(
    data = ref_arrow,
    aes(x = heures + u + 0.5, y = v + vv, label = label),
    inherit.aes = FALSE,
    hjust = 0,
    vjust = 0.5,
    size = 3,
    color = "black"
  ) +
  scale_color_manual(values = c("Météo France" = "#619CFF", "ERA5" = "#F8766D", "ERA5-Land" = "#00BA38")) +
  labs(color = "Source")
```

```{r, eval = savefig}
# plot coloré AVEC BACKGROUND WHITE
ggsave(
  filename = paste0(figpath, "comparison_MAR_wind_3j_colors.png"),
  width = 12, height = 6, dpi = 600, bg = "white"
)
rm(df_3j_long)
```
```{r}
# roses des vents

df_long <- df %>%
  select(date, starts_with("ws"), starts_with("wd")) %>%
  pivot_longer(
    cols = -date,
    names_to = c(".value", "source"),
    names_pattern = "^(ws|wd)_(.*)$"
  ) %>%
  filter(!is.na(ws), !is.na(wd)) %>%
  mutate(
    source = recode(source, 
                    "mf" = "Météo France", 
                    "era5" = "ERA5", 
                    "era5land" = "ERA5-Land")
  )
windrose <- windRose(df_long, type = "source", paddle = FALSE, border = TRUE,
                     key.header = "Source", key.footer = "Vitesse du vent (m/s)",
                     breaks = c(0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20))
```

```{r}
# on groupe par heure de la journée et on calcule le biais MF/ERA5 et MF/ERA5 land pour chaque heure pour le vent, et on plot
df_hourly <- df %>%
  mutate(
    hour = hour(date),
    err_u_era5 = windu_mf - windu_era5,
    err_v_era5 = windv_mf - windv_era5,
    err_u_era5land = windu_mf - windu_era5land,
    err_v_era5land = windv_mf - windv_era5land
  ) %>%
  group_by(hour) %>%
  summarise(
    windu_mf = mean(windu_mf, na.rm = TRUE),
    windu_era5 = mean(windu_era5, na.rm = TRUE),
    windu_era5land = mean(windu_era5land, na.rm = TRUE),
    windv_mf = mean(windv_mf, na.rm = TRUE),
    windv_era5 = mean(windv_era5, na.rm = TRUE),
    windv_era5land = mean(windv_era5land, na.rm = TRUE),
    
    rmse_u_era5 = sqrt(mean(err_u_era5^2, na.rm = TRUE)),
    rmse_v_era5 = sqrt(mean(err_v_era5^2, na.rm = TRUE)),
    rmse_u_era5land = sqrt(mean(err_u_era5land^2, na.rm = TRUE)),
    rmse_v_era5land = sqrt(mean(err_v_era5land^2, na.rm = TRUE))
  ) %>%
  mutate(
    bias_windu_era5 = windu_mf - windu_era5,
    bias_windv_era5 = windv_mf - windv_era5,
    bias_windu_era5land = windu_mf - windu_era5land,
    bias_windv_era5land = windv_mf - windv_era5land,
    
    bias_ws_era5 = sqrt(bias_windu_era5^2 + bias_windv_era5^2),
    bias_ws_era5land = sqrt(bias_windu_era5land^2 + bias_windv_era5land^2),
    bias_wd_era5 = atan2(bias_windv_era5, bias_windu_era5) * 180 / pi,
    bias_wd_era5land = atan2(bias_windv_era5land, bias_windu_era5land) * 180 / pi,
    
    rmse_ws_era5 = sqrt(rmse_u_era5^2 + rmse_v_era5^2),
    rmse_ws_era5land = sqrt(rmse_u_era5land^2 + rmse_v_era5land^2)
  )

#Biais vitesse du vent 
a <- ggplot(df_hourly, aes(x = hour)) +
  geom_line(aes(y = bias_ws_era5, color = "ERA5")) +
  geom_line(aes(y = bias_ws_era5land, color = "ERA5-Land")) +
  labs(x = "Heure de la journée", y = "Biais de la vitesse du vent (m/s)", color = "Source") +
  scale_x_continuous(breaks = 0:23) +
  scale_color_manual(values = c("ERA5" = "#F8766D", "ERA5-Land" = "#00BA38")) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 15, hjust = 1))

# direction du vent 
b <- ggplot(df_hourly, aes(x = hour)) +
  geom_line(aes(y = bias_wd_era5, color = "ERA5")) +
  geom_line(aes(y = bias_wd_era5land, color = "ERA5-Land")) +
  labs(x = "Heure de la journée", y = "Biais de la direction du vent (°)", color = "Source") +
  scale_x_continuous(breaks = 0:23) +
  scale_color_manual(values = c("ERA5" = "#F8766D", "ERA5-Land" = "#00BA38")) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 15, hjust = 1))

# RMSE vitesse du vent
c <- ggplot(df_hourly, aes(x = hour)) +
  geom_line(aes(y = rmse_ws_era5, color = "ERA5")) +
  geom_line(aes(y = rmse_ws_era5land, color = "ERA5-Land")) +
  labs(x = "Heure de la journée", y = "RMSE de la vitesse du vent (m/s)", color = "Source") +
  scale_x_continuous(breaks = 0:23) +
  scale_color_manual(values = c("ERA5" = "#F8766D", "ERA5-Land" = "#00BA38")) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 15, hjust = 1))

print(a)
print(b)
print(c)
```
```{r, eval = savefig}
ggsave(
  filename = paste0(figpath, "comparison_MAR_windspeed_bias.png"),
  plot = a,
  width = 18, height = 10, dpi = 300, units = "cm", bg = "white"
)

ggsave(
  filename = paste0(figpath, "comparison_MAR_winddir_bias.png"),
  plot = b,
  width = 18, height = 10, dpi = 600, bg = "white", units = "cm"
)

ggsave(
  filename = paste0(figpath, "comparison_MAR_windspeed_rmse.png"),
  plot = c,
  width = 18, height = 10, dpi = 600, bg = "white", units = "cm")
```


# Comparaison des données pour la température

```{r}
ggplot(df_3j, aes(x = date)) +
  geom_line(aes(y = t2m_mf, color = "Météo France")) +
  geom_line(aes(y = t2m_era5, color = "ERA 5")) +
  geom_line(aes(y = t2m_era5land, color = "ERA5 land")) +
  labs(x = "Date", y = "Temperature (°C)", color = "Source") +
  scale_x_datetime(date_labels = "%Y-%m-%d %H:%M", date_breaks = "1 day") +
  scale_color_manual(values = c("Météo France" = "#619CFF", "ERA 5" = "#F8766D", "ERA5 land" = "#00BA38")) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(angle = 15, hjust = 1)
  )
```

```{r, eval = savefig}
ggsave(
  filename = paste0(figpath, "comparison_MAR_temp_3j.png"),
  width = 12, height = 6, dpi = 600, bg = "white"
)
```
```{r}
# rmse et biais par heure de la journée

# Calcul du biais et de la RMSE pour t2m
df_hourly_t2m <- df %>%
  mutate(
    hour = hour(date),
    err_t2m_era5 = t2m_mf - t2m_era5,
    err_t2m_era5land = t2m_mf - t2m_era5land
  ) %>%
  group_by(hour) %>%
  summarise(
    t2m_mf = mean(t2m_mf, na.rm = TRUE),
    t2m_era5 = mean(t2m_era5, na.rm = TRUE),
    t2m_era5land = mean(t2m_era5land, na.rm = TRUE),
    
    rmse_t2m_era5 = sqrt(mean(err_t2m_era5^2, na.rm = TRUE)),
    rmse_t2m_era5land = sqrt(mean(err_t2m_era5land^2, na.rm = TRUE))
  ) %>%
  mutate(
    bias_t2m_era5 = t2m_mf - t2m_era5,
    bias_t2m_era5land = t2m_mf - t2m_era5land
  )

g1 <- ggplot(df_hourly_t2m, aes(x = hour)) +
  geom_line(aes(y = bias_t2m_era5, color = "ERA5")) +
  geom_line(aes(y = bias_t2m_era5land, color = "ERA5-Land")) +
  labs(x = "Heure de la journée", y = "Biais de température (°C)", color = "Source") +
  scale_x_continuous(breaks = 0:23) +
  scale_color_manual(values = c("ERA5" = "#F8766D", "ERA5-Land" = "#00BA38")) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 15, hjust = 1))

g2 <- ggplot(df_hourly_t2m, aes(x = hour)) +
  geom_line(aes(y = rmse_t2m_era5, color = "ERA5")) +
  geom_line(aes(y = rmse_t2m_era5land, color = "ERA5-Land")) +
  labs(x = "Heure de la journée", y = "RMSE température (°C)", color = "Source") +
  scale_x_continuous(breaks = 0:23) +
  scale_color_manual(values = c("ERA5" = "#F8766D", "ERA5-Land" = "#00BA38")) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 15, hjust = 1))

print(g1)
print(g2)

```


# Comparaison des données pour les précipitations
```{r}
# on fait un histogramme du cumul par jour par source
df_precip <- df %>%
  select(date, totalprecipitation_mf, totalprecipitation_era5) %>%
  mutate(date = as.Date(date)) %>%
  group_by(date) %>%
  summarise(
    MeteoFrance = sum(totalprecipitation_mf, na.rm = TRUE),
    ERA5 = sum(totalprecipitation_era5, na.rm = TRUE)
  ) %>%
  pivot_longer(cols = -date, names_to = "source", values_to = "precipitation")

# on fait un graphique
ggplot(df_precip, aes(x = date, y = precipitation, fill = source)) +
  geom_bar(stat = "identity", position = "dodge") +
  labs(x = "Date", y = "Cumul de précipitations (mm)", fill = "Source") +
  scale_x_date(date_labels = "%Y-%m-%d", date_breaks = "1 month") +
  scale_fill_manual(values = c("MeteoFrance" = "#619CFF", "ERA5" = "#F8766D")) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(angle = 15, hjust = 1)
  )
```

```{r}
ggsave(
  filename = paste0(figpath, "comparison_MAR_precip.png"),
  width = 12, height = 6, dpi = 600, bg = "white"
)
```


```{r}
# on fait un histogramme genre jours sans pluie/jours à moins de 5 mm/blabla
hist_precip <- df %>%
  select(date, totalprecipitation_mf, totalprecipitation_era5) %>%
  mutate(date = as.Date(date)) %>%
  group_by(date) %>%
  summarise(
    MeteoFrance = sum(totalprecipitation_mf, na.rm = TRUE),
    ERA5 = sum(totalprecipitation_era5, na.rm = TRUE)
  ) %>%
  pivot_longer(cols = -date, names_to = "source", values_to = "precipitation") %>%
  mutate(
    category = case_when(
      precipitation == 0 ~ "Pas de précipitations",
      precipitation > 0 & precipitation <= 3 ~ "Précipitations < 3mm",
      precipitation > 3 & precipitation <= 5 ~ "Précipitations entre 3 et 5mm",
      precipitation > 5 ~ "Précipitations > 5mm"
    )
  )
ggplot(hist_precip, aes(x = category, fill = source)) +
  geom_bar(position = "dodge") +
  labs(x = "", y = "Nombre de jours", fill = "Source") +
  scale_fill_manual(values = c("MeteoFrance" = "#619CFF", "ERA5" = "#F8766D")) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(angle = 0, hjust = 1)
  )
```
```{r, eval = savefig}
ggsave(
  filename = paste0(figpath, "comparison_MAR_histprecip.png"),
  width = 12, height = 8, dpi = 600, bg = "white"
)
```
# Comparaison des données pour le rayonnement solaire
```{r}
ggplot(df_3j, aes(x = date)) +
  geom_line(aes(y = ssrd_mf, color = "Météo France")) +
  geom_line(aes(y = ssrd_era5land, color = "ERA5-Land")) +
  labs(x = "Date", y = "Rayonnement global (W/m²)", color = "Source") +
  scale_x_datetime(date_labels = "%Y-%m-%d %H:%M", date_breaks = "1 day") +
  scale_color_manual(values = c("Météo France" = "#619CFF", "ERA5-Land" = "#00BA38")) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(angle = 15, hjust = 1)
  )
```
